server/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
eventlet==0.36.1
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
Werkzeug==3.0.3
